Add Composer scripts for code sniffing and autofixing
There now are the following new Composer scripts:

- `ci`: run all CI-related Composer scripts
- `ci:static` run the static code analysis
- `phpcs` run PHP_CodeSniffer
- `fix` fix the autofixable PHP code warnings
- `phpcbf` fix the autofixable PHP_CodeSniffer warnings

For starters, we have a sniff that checks that all files are in
PHP strict mode. (The CI builds do not run this check as the files
are not strict yet at this point.)

The PHP_CodeSniffer autofixer cannot make files strict, though
(as this is a non-autofixable code sniff). It will be able to
autofix other things we'll add later, though.